Expand log view functions
I am not sure what the "correct" order is, but this is the order in
which I would want them so that I can do an verify-add-verify cycle with
incoming lines.

It's also worth considering optional verification. The whole API is not
particularly well-considered at this point, so we'll see.