Try to maintain our filesystem consistency gaurantee
If we try to read valid stuff off the disk and its not, we remove it
right there.

None of this is atomic (and stuff can change on disk outside of here) so
it's possible to send bad data elsewhere if it's not verified on the way
out.

Expectation is that the other side verifies before acceptng, so we only
hurt ourselves.