Add in options parsing
This is part of an effort to improve consumer ergonomics.  The intent is
that they never have to deal with anything outside of the main `Baobab`
module.

So we need to be able to set the options to pass everything around.
It also involves removing a couple internal items from the public API