Wrap a bit more tightly around Pockets
Still fighting the fact that there is not an application here, so we do
a lot of opening and closing.  I may well switch to dets because I can do
the matching via their API instead of streaming all keys and such
(And I cannot even stream the keys anymore because I closed the file
whence it comes)

This simplification may be a step on that road.

Also moved to `:contents` for the file contents request.  `:content`
iwas getting too overloaded due to m limited vocabulary.