Sanitise persistence interface a bit
The ergonomics still kinda suck, but we now trust that what we put into
the store is the same thing we will take out.  We do, of course, provide
a long-standing way to revalidate if those assumptions get broken.

The biggest difference is the the appending to logs where we trust
ourselves to have built it corectly and do not reverify all of our
asusmptions.  We'd get them wrong yet again, anyway.