Hit disk less often
On store, presume that what we put in is what hit persistence and will
be returned.  So we can validate it up front then only store when it is
good.

This also involves doing the signature calc on the Entry contents. Which
is also less work!

Take a step toward more sane file handling